Fix merkle tree (#78)
This PR fixes and enables the merkle tree implementation in quaireaux.

Rebased on top of main with alpha.7 merged, so works with alpha.7

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Merkle tree implementation is not working nor ready because the core of
it is commented out. This was previously because of a bug in Cairo, but
that bug is fixed now.

## What is the new behavior?

Merkle tree is now usable and has working tests.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->