fix: integer overflow in fast_power (#104)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: #95 

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- panic with overflow error for certain inputs

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->

## Other information

<!-- Any other information that is important to this PR, such as
screenshots of how the component looks before and after the change. -->

I have choose to change all the parameter types to `u256` instead of
using
[`u128_wide_mul`](https://github.com/starkware-libs/cairo/blob/77ae28f996c0960ce5cfc926703f60bae8d5db5a/corelib/src/integer.cairo#L66C1-L69).

Because in the loop:
- we would need to do it in 2 places, so 2 type conversions to u256
- `%` is only implemented for same types we cannot calculate `u256 %
u128`, so would need to convert `modulus` to u256
- after calculating modulus would needs to cast it back to u128

Also for some reason comment mentioned result of function is `felt252`
instead of `u128`, so have fixed that as well.