fix(encoding/base64): use character set passed as argument (#112)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

`encode_inner` is already passed a character set, it should use that
instead of generating it again. This will cause `Base64Encoder` to also
use `-` and `_` in character set when it should have been `+` and `/`.

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Now character set are properly used
- Also added few unit test to make sure regressions don't happen

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->

## Other information

<!-- Any other information that is important to this PR, such as
screenshots of how the component looks before and after the change. -->

Used https://cryptii.com/pipes/binary-to-base64 to calculate result for
adding unit tests.