feat(storage): list (#127)
<!--- Please provide a general summary of your changes in the title
above -->

Adding support for array storage (up to 2^32 elements). This feature
enables things like:

```rust
#[storage]
struct Storage {
    values: List<u256>
}
```

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Currently, when you want to store an array of values in a Starknet
contracts, you have to resort to sth like this

```rust
struct Storage {
    values_count: u32,
    values: LegacyMap<u32, u256>
}
```

i.e. have a dictionary where the key is the index of the element. This
is cumbersome as you have to take care of properly updating both values
and just deal with multiple storage variables all the time.

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

This PR adds a new data structure called `List` which rectifies the
above approach. There's only one storage variable and in general it acts
as an "array" from other programming languages that devs are used to (or
rather it tries its best to do so).

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->

## Other information

The TL;DR of the implementation is that the storage segment where the
List values are written is based on the name of the storage variable and
a index, hashed together. I've written about it in more detail in the
inline comments, so I won't reproduce it here, but please feel free to
ask any questions.

Storing most (all?) corelib values in a List should work out of the box.
For custom types, devs will have to implement a simple trait
`StorageSize` - the reason why is also mentioned in the inline comment
in the code.

Using `append` costs 2 storage writes, but that's the same as with the
counter + map approach described above.

<!-- Any other information that is important to this PR, such as
screenshots of how the component looks before and after the change. -->

I'm happy to hear any kind of feedback about the implementation. It's
also completely fine if you deem this not be a good fit in Alexandria, I
can just publish it on my own.

Beyond that, I have a list of questions I'd like to hear yours'
opinions;

### Open questions:

1.  Is the structure - creating a new module called `storage` ok?
2. ~~Should `get` return an `Option<T>`? In other words, if getting an
element at index > len, should the return value be `Option::None`? That
would be similar to corelib's array.~~ UPDATE: `get` now returns
`Option<T>`.
3. Should there be a `span` fn that returns a `Span<T>` of all the
elements in the list?
4. Should it use Poseidon for hashing rather than `LegacyHash`
(Pedersen)?
5. Should it use `usize` instead of `u32` for `len` and `index`?
Currently, `usize` is an alias for `u32` in Cairo, but that might change
it the future. What's the best option?
6. ~~Should there be a `pop_front` function that removes the "topmost"
element?~~ UPDATE: there's now a `pop_front` returning `Option<T>`.

I think this is also the first Starknet specific module in Alexandria.
As such, it needs the `--starknet` flag to be passed with `cairo-test`.
How should this be done? A solution I though of was to have two targets,
`test-cairo` and `test-starknet` and make them dependencies of `test`,
so sth like

```makefile
test: test-cairo test-starknet

test-cairo: test-data_structures test-encoding test-math test-sorting test-searching

test-starknet: test-storage

#...

test-storage:
	@echo "Testing storage"
	cairo-test --starknet $(PROJECT_NAME)/storage
```

But I'm far from an expert in Makefiles 😁

Also, as discussed in the TG group, this works only with Cairo v2 due to
the new `StorageAccess` trait (it doesn't have the internal functions in
v1.1).

---------

Co-authored-by: Lucas @ StarkWare <70894690+LucasLvy@users.noreply.github.com>