dev: test everything in Makefile (#135)
Follow up on #127, adds testing of `storage` to the Makefile's `test`
target.

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [x] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

When running `make test`, the `test-storage` was not run. It is the
first target/module that requires a `--starknet` arg to `cairo-test`, so
it couldn't have been done like the rest of the stuff.

Issue Number: N/A

## What is the new behavior?

Running `make test` runs the tests for `storage` as well.

There's also a 2 new targets: `test-cairo` and `test-starknet`.
They group other targets as dependencies, former one is intended for
non-starknet modules while the latter one for those that are starknet
specific.

<!-- Please describe the behavior or changes that are being added by
this PR. -->

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->

## Other information

<!-- Any other information that is important to this PR, such as
screenshots of how the component looks before and after the change. -->

I'm no a Makefile expert by any stretch of the imagination so I'm open
to any suggestions how this can be made better.