refactor: group packages (#136)
## Pull Request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [X] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Library is currently divided in multiple scarb packages. 
Related to discussion on issue #130 

## What is the new behavior?

- Group all packages into one
- Remove circular dependencies

## Does this introduce a breaking change?

- [X] Yes
- [ ] No

As @gaetbout suggested we can create a tag that people can use to
reference this repo "pre-package merge"

## Other information

<!-- Any other information that is important to this PR, such as
screenshots of how the component looks before and after the change. -->

---------

Co-authored-by: Lucas @ StarkWare <70894690+LucasLvy@users.noreply.github.com>