feat: from_array for List (#183)
## Pull Request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Issue Number:
[#158](https://github.com/keep-starknet-strange/alexandria/issues/158)

## What is the new behavior?

- Added the `from_array` method for `ListTrait`
- Added the corresponding unit tests

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

## Other information

The implementation is very similar to the `append` function except that
it sets the length only once.

There's one question still, what should happen when you call the
function with an empty list, like
```rust
lst.from_array(array![]);
```
For my current implementation it "cleans" the list (basically set the
length to zero) but maybe this is not what we want. Let me know! 😁

---------

Co-authored-by: akhercha <akherchache@pm.me>