dev: Gas metering  (#319)
<!--- Please provide a general summary of your changes in the title
above -->

Adds a gas_report for the PRs in Github Actions!
If the gas increases, the PR check will fail. If the gas decreases or
remains same, the PR check passes!

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [x] Other (please describe): Github Action job 

## What is the current behavior?

There is no tracing of gas usage in PRs!
Issue Number: #292 

## What is the new behavior?

When the PR is made, the gas_report is generated and compared with the
pre-existing one to trace the gas usage in tests. If more gas is being
used, the PR check fails, and has to be manually accepted after
checking! If the gas usage remains same, or decreases, the PR passes!

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->