feat(macros): add sub mul div zero derives (#328)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):


## What is the new behavior?

PR introduces a bunch of new macros that allow to derive `Add`, `Sub`,
`Mul`, `Div`, `AddAssign`, `SubAssign`, `MulAssign`, `DivAssign` and
`Zero` traits. Now we can have a struct like this:

```
#[derive(Add, AddAssign, Sub, SubAssign, Mul, MulAssign, Div, DivAssign, Zero)]
struct Foo {
    x: u8,
    y: u16,
    z: u32
}
```

and all the basic ops are available out of the box.

The macros work for more complex types as well (generic & nested
structs).

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

## Other information

<!-- Any other information that is important to this PR, such as
screenshots of how the component looks before and after the change. -->

I've cleaned up the package structure to be more modular as it grows.