fix: byte array ABI encoding (#347)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

When (ABI) encoding a ByteArray or Bytes with size multiple of 32, we
try to add an unnecessary 32 byte padding and end up with the following
panic:

> Panicked with 0x496e646578206f7574206f6620626f756e6473 ('Index out of
bounds').

Issue Number: N/A

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

If the size is multiple of 32, skip padding. This avoids the panic.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->

## Other information

N/A

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>