Adds a public function to generate a bitcoin address from public key hash (#413)
* Refactor relevant function to take a pubkey_hash as input
* Add tests for pubkey_hash to address
* Fix some typos

<!--- Please provide a general summary of your changes in the title
above -->

## Pull Request type

<!-- Please try to limit your pull request to one type; submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [X] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no API changes)
- [ ] Build-related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Currenlty it is only possible to generat an address from a private key
or a public key
Issue Number: #412 

## What is the new behavior?

A new public function that takes a public key hash and an address type
as parameters

Returns a bitcoin address if creating one from pubkey hash is possible

Panics if the address type is not compatible with address generation
from pubkey hash



## Does this introduce a breaking change?

- [ ] Yes
- [X] No

<!-- If this does introduce a breaking change, please describe the
impact and migration path for existing applications below. -->

## Other information

This adds a new API function, existing ones are not changed in
functionality